third_party/blink/web_tests/platform/generic/fast/forms/output/htmloutputelement-reset-event-expected.txt

A Test for sending a reset event to output elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

- Sets the value attribute of the output element.
PASS output.defaultValue is "defaultValue"
PASS output.value is "aValue"
PASS output.innerText is "aValue"
PASS output.innerHTML is "aValue"
- Sends a reset event to reset the value to the default value.
PASS output.defaultValue is "defaultValue"
PASS output.value is "defaultValue"
PASS output.innerText is "defaultValue"
PASS output.innerHTML is "defaultValue"
- Ensures that the value mode flags is in mode "default".
PASS output.defaultValue is "another defaultValue"
PASS output.value is "another defaultValue"
PASS output.innerText is "another defaultValue"
PASS output.innerHTML is "another defaultValue"
PASS successfullyParsed is true

TEST COMPLETE
